refactor(tests): split oversized test files
Split test_csp_header.py (370L) into policy/webview_token/host_header
siblings plus a shared _csp_helpers.py.
Split test_llm_bridge_routes.py (404L) into base/omlx siblings.
Split test_rate_limit_hardening.py (433L) into base/cross_process_lock
siblings.
Split test_server_coverage.py (305L) into base/serve_modes siblings.
Split test_omlx.py (332L) into base/models siblings.
Split test_cancel_discard_purge_state.py (382L) into base/paths siblings.

develop picked up a size ratchet covering tests/ at file level
(dd083892) after this branch's last rebase; these six grew past the
300-line cap from this cycle's own test additions.
